Fix Windows warnnings.

+++ modules/xfeatures2d/src/brief.cpp
@@ -91,8 +91,8 @@ inline int smoothedSum(const Mat& sum, const KeyPoint& pt, int y, int x, bool us
 
     if ( use_orientation )
     {
-      int rx = ((float)x)*R(1,0) - ((float)y)*R(0,0);
-      int ry = ((float)x)*R(0,0) + ((float)y)*R(1,0);
+      int rx = (int)(((float)x)*R(1,0) - ((float)y)*R(0,0));
+      int ry = (int)(((float)x)*R(0,0) + ((float)y)*R(1,0));
       if (rx > 24) rx = 24; if (rx < -24) rx = -24;
       if (ry > 24) ry = 24; if (ry < -24) ry = -24;
       x = rx; y = ry;
